cleaning the backend code
